update; removed auto gen IDs from tables and entities

+++ src/main/java/com/mdsujan/restPostgres/service/ItemService.java
@@ -3,6 +3,7 @@ package com.mdsujan.restPostgres.service;
 import com.mdsujan.restPostgres.entity.Item;
 import com.mdsujan.restPostgres.repository.DemandRepository;
 import com.mdsujan.restPostgres.repository.ItemRepository;
+import com.mdsujan.restPostgres.repository.LocationRepository;
 import com.mdsujan.restPostgres.repository.SupplyRepository;
 import com.mdsujan.restPostgres.request.CreateItemRequest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,8 +14,8 @@ import java.util.List;
 @Service
 public class ItemService {
     // While constructing the ItemService object, if there's no constructor or setter method
-    // to inject the ItemRepository bean,
-    // the container will use reflection to inject ItemRepository into ItemService.
+    // to inject the ItemRepository bean, the container will use reflection
+    // to inject ItemRepository into ItemService.
     @Autowired
     ItemRepository itemRepository;
 
@@ -23,6 +24,7 @@ public class ItemService {
 
     @Autowired
     DemandRepository demandRepository;
+
     public List<Item> getAllItems() {
         return itemRepository.findAll();
     }
@@ -33,24 +35,27 @@ public class ItemService {
     }
 
     public Item createItem(CreateItemRequest createItemRequest) {
-        Item newItem = new Item(createItemRequest);
         // new record should not be created if record already exists
-        // if record does not exist then we persist into the table item
-        if(itemRepository.findById(createItemRequest.getId()).get() == null){
-            itemRepository.save(newItem);
-            return newItem;
+
+        // if record with same id exists then simply return it
+        if (itemRepository.findById(createItemRequest.getId()).isPresent()) {
+            return itemRepository.findById(createItemRequest.getId()).get();
         }
-        // else return the old record
-        return itemRepository.findById(createItemRequest.getId()).get();
+        // else we create a new item
+        Item item = new Item(createItemRequest);
+        itemRepository.save(item);
+        return item;
     }
 
     public boolean deleteItemById(Long itemId) {
         try {
+            if (supplyRepository.findById(itemId).isPresent() && demandRepository.findById(itemId).isPresent()) {
+                return false;
+            }
             // delete only if no supply or demand exists for this item
-
             itemRepository.deleteById(itemId);
             return true;
-        }catch (Exception e){
+        } catch (Exception e) {
             return false;
         }
     }
